80-finishing-the-chat-bot-ui
- Structure changed
- code not needed deleted
- chatbot name added
- keyboard bug fixed
