+++ android/app/src/main/java/at/sozialministerium/youngcarers/Navigation.kt
@@ -102,13 +102,6 @@ fun Navigation(
             OnBoarding(navController = navController)
 
         }
-        // Onboardscreen
-        composable(Welcome.Home.route){
-            HomeScreen()
-
-        }
-
-
     }
 }
